refactor: put menu data to new script
