+++ pages/superadmin/clubs/index.js
@@ -1,51 +1,62 @@
 import Head from "next/head";
-import Image from "next/image";
-import Link from "next/link";
-import styles from "../../../styles/Home.module.css";
+import MaterialTable from "material-table";
+import { useStyles } from "../../../styles/clubs/Index.style";
+import { TopBar } from "../../../src/components/navigation/TopBar";
+import { MenuBar } from "../../../src/components/navigation/MenuBar";
 
-export default function ClubsSuperAdmin() {
+export default function DashboardSuperAdmin() {
+  const classes = useStyles();
+
   return (
-    <div className={styles.container}>
+    <div className={classes.root}>
       <Head>
-        <title>Calfit Clubs SuperAdmin</title>
+        <title>Dashboard CalFit</title>
         <meta name="description" content="Generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
 
-      <main className={styles.main}>
-        <h1>Clubs</h1>
-        <Link href="/superadmin/dashboard" passHref>
-          <div className={styles.newsdetail}>SuperAdmin Dashboard</div>
-        </Link>
-        <Link href="/superadmin/classes" passHref>
-          <div className={styles.newsdetail}>SuperAdmin Classes Access</div>
-        </Link>
-        <Link href="/superadmin/newsletters" passHref>
-          <div className={styles.newsdetail}>SuperAdmin Newsletters Access</div>
-        </Link>
-        <Link href="/superadmin/videos" passHref>
-          <div className={styles.newsdetail}>SuperAdmin Videos Access</div>
-        </Link>
-        <Link href="/superadmin/admins" passHref>
-          <div className={styles.newsdetail}>SuperAdmin Admins Access</div>
-        </Link>
-        <Link href="/superadmin/profile" passHref>
-          <div className={styles.newsdetail}>SuperAdmin Profile Settings</div>
-        </Link>
-      </main>
+      <TopBar label="Dashboard" />
 
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{" "}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
+      <main className={classes.main}>
+        <MenuBar selected={"Clubs"} />
+        <MaterialTable
+          title="Positioning Actions Column Preview"
+          columns={[
+            { title: 'Name', field: 'name' },
+            { title: 'Surname', field: 'surname' },
+            { title: 'Birth Year', field: 'birthYear', type: 'numeric' },
+            { title: 'Birth Year', field: 'birthYear', type: 'numeric' },
+            { title: 'Birth Year', field: 'birthYear', type: 'numeric' },
+            { title: 'Birth Year', field: 'birthYear', type: 'numeric' },
+            {
+              title: 'Birth Place',
+              field: 'birthCity',
+              lookup: { 34: 'İstanbul', 63: 'Şanliurfa' },
+            },
+          ]}
+          data={[
+            { name: 'Mehmet', surname: 'Baran', birthYear: 1987, birthCity: 63 },
+            { name: 'Zerya Betül', surname: 'Baran', birthYear: 2017, birthCity: 34 },
+          ]}
+          actions={[
+            {
+              icon: 'save',
+              tooltip: 'Save User',
+              onClick: (event, rowData) => alert("You saved " + rowData.name)
+            },
+            rowData => ({
+              icon: 'delete',
+              tooltip: 'Delete User',
+              onClick: (event, rowData) => confirm("You want to delete " + rowData.name),
+              disabled: rowData.birthYear < 2000
+            })
+          ]}
+          options={{
+            actionsColumnIndex: -1
+          }}
+        />
+
+      </main>
     </div>
   );
 }
